pyproject.toml
[project]
name = "thinkfeelshare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-socketio>=5.13.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "alembic>=1.16.1",
]
